Move CommitTarget to _utility.py
