+++ backup_cosense/_git.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import dataclasses
 import datetime
-import itertools
 import json
 import logging
 import os
@@ -17,7 +16,8 @@ import jsonschema
 
 from ._backup import BackupInfoJSON, jsonschema_backup_info
 from ._json import parse_json
-from .exceptions import CommitTargetError, GitNotFoundError
+from ._utility import CommitTarget
+from .exceptions import GitNotFoundError
 
 
 @dataclasses.dataclass(frozen=True)
@@ -66,42 +66,6 @@ class Commit:
         if not body:
             return header
         return "\n".join([header, "", *body])
-
-
-@dataclasses.dataclass
-class CommitTarget:
-    added: set[pathlib.Path] = dataclasses.field(default_factory=set)
-    updated: set[pathlib.Path] = dataclasses.field(default_factory=set)
-    deleted: set[pathlib.Path] = dataclasses.field(default_factory=set)
-
-    def __post_init__(self) -> None:
-        self.normalize()
-        self.validate()
-
-    def normalize(self) -> None:
-        self.added = {path.resolve() for path in self.added}
-        self.updated = {path.resolve() for path in self.updated}
-        self.deleted = {path.resolve() for path in self.deleted}
-
-    def validate(self) -> None:
-        errors: list[str] = []
-        # check intersection
-        for set1, set2 in itertools.combinations(["added", "updated", "deleted"], 2):
-            for path in getattr(self, set1) & getattr(self, set2):
-                errors.append(f'"{path}" exists in both "{set1}" and "{set2}"')
-        # raise error
-        if errors:
-            raise CommitTargetError("".join(errors))
-
-    def update(self, other: CommitTarget) -> CommitTarget:
-        self.added |= other.added
-        self.updated |= other.updated
-        self.deleted |= other.deleted
-        self.validate()
-        return self
-
-    def is_empty(self) -> bool:
-        return not (self.added or self.updated or self.deleted)
 
 
 class Git:
